services/renderer-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
Pillow==10.2.0
numpy==1.26.3
opencv-python-headless==4.9.0.80
redis==5.0.1
aioredis==2.0.1
pydantic==2.5.3
python-multipart==0.0.6

